pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brui_core"
version = "2.0.0"
description = "Core browser UI automation framework"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Ryan Zheng", email = "ryan.zheng.work@gmail.com" }
]
dependencies = [
  "playwright==1.55.0",
  "pyperclip",
  "Pillow",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Testing",
]

[project.optional-dependencies]
test = [
  "pytest-playwright==0.4.4",
  "playwright==1.55.0",
  "pytest-asyncio",
]

[project.urls]
Homepage = "https://github.com/AutoByteus/brui_core.git"
Source = "https://github.com/AutoByteus/brui_core.git"

[tool.setuptools.packages.find]
include = ["brui_core", "brui_core.*"]
